requirements.txt
marshmallow
python-igraph
